Cleanup core package
License: MIT
