+++ core/builder.go
@@ -5,57 +5,24 @@ import (
 	"crypto/rand"
 	"encoding/base64"
 	"errors"
-	"os"
-	"syscall"
 
 	"go.uber.org/fx"
 
-	"github.com/ipfs/go-ipfs/p2p"
-	"github.com/ipfs/go-ipfs/provider"
+	"github.com/ipfs/go-ipfs/core/bootstrap"
+	"github.com/ipfs/go-ipfs/core/node"
 
 	repo "github.com/ipfs/go-ipfs/repo"
 
 	ds "github.com/ipfs/go-datastore"
 	dsync "github.com/ipfs/go-datastore/sync"
 	cfg "github.com/ipfs/go-ipfs-config"
-	offline "github.com/ipfs/go-ipfs-exchange-offline"
-	offroute "github.com/ipfs/go-ipfs-routing/offline"
 	metrics "github.com/ipfs/go-metrics-interface"
 	resolver "github.com/ipfs/go-path/resolver"
 	ci "github.com/libp2p/go-libp2p-crypto"
 	peer "github.com/libp2p/go-libp2p-peer"
 )
 
-type BuildCfg struct {
-	// If online is set, the node will have networking enabled
-	Online bool
-
-	// ExtraOpts is a map of extra options used to configure the ipfs nodes creation
-	ExtraOpts map[string]bool
-
-	// If permanent then node should run more expensive processes
-	// that will improve performance in long run
-	Permanent bool
-
-	// DisableEncryptedConnections disables connection encryption *entirely*.
-	// DO NOT SET THIS UNLESS YOU'RE TESTING.
-	DisableEncryptedConnections bool
-
-	// If NilRepo is set, a Repo backed by a nil datastore will be constructed
-	NilRepo bool
-
-	Routing RoutingOption
-	Host    HostOption
-	Repo    repo.Repo
-}
-
-func (cfg *BuildCfg) getOpt(key string) bool {
-	if cfg.ExtraOpts == nil {
-		return false
-	}
-
-	return cfg.ExtraOpts[key]
-}
+type BuildCfg node.BuildCfg
 
 func (cfg *BuildCfg) fillDefaults() error {
 	if cfg.Repo != nil && cfg.NilRepo {
@@ -77,11 +44,11 @@ func (cfg *BuildCfg) fillDefaults() error {
 	}
 
 	if cfg.Routing == nil {
-		cfg.Routing = DHTOption
+		cfg.Routing = node.DHTOption
 	}
 
 	if cfg.Host == nil {
-		cfg.Host = DefaultHostOption
+		cfg.Host = node.DefaultHostOption
 	}
 
 	return nil
@@ -115,8 +82,6 @@ func defaultRepo(dstore repo.Datastore) (repo.Repo, error) {
 	}, nil
 }
 
-type MetricsCtx context.Context
-
 // NewNode constructs and returns an IpfsNode using the given cfg.
 func NewNode(ctx context.Context, cfg *BuildCfg) (*IpfsNode, error) {
 	if cfg == nil {
@@ -141,12 +106,12 @@ func NewNode(ctx context.Context, cfg *BuildCfg) (*IpfsNode, error) {
 	})
 
 	// TODO: Remove this, use only for passing node config
-	cfgOption := fx.Provide(func() *BuildCfg {
-		return cfg
+	cfgOption := fx.Provide(func() *node.BuildCfg {
+		return (*node.BuildCfg)(cfg)
 	})
 
-	metricsCtx := fx.Provide(func() MetricsCtx {
-		return MetricsCtx(ctx)
+	metricsCtx := fx.Provide(func() node.MetricsCtx {
+		return node.MetricsCtx(ctx)
 	})
 
 	params := fx.Options(
@@ -155,58 +120,12 @@ func NewNode(ctx context.Context, cfg *BuildCfg) (*IpfsNode, error) {
 		metricsCtx,
 	)
 
-	storage := fx.Options(
-		fx.Provide(repoConfig),
-		fx.Provide(datastoreCtor),
-		fx.Provide(baseBlockstoreCtor),
-		fx.Provide(gcBlockstoreCtor),
-	)
-
-	ident := fx.Options(
-		fx.Provide(identity),
-		fx.Provide(privateKey),
-		fx.Provide(peerstore),
-	)
-
-	ipns := fx.Options(
-		fx.Provide(recordValidator),
-	)
-
-	providers := fx.Options(
-		fx.Provide(providerQueue),
-		fx.Provide(providerCtor),
-		fx.Provide(reproviderCtor),
-
-		fx.Invoke(reprovider),
-		fx.Invoke(provider.Provider.Run),
-	)
-
-	online := fx.Options(
-		fx.Provide(onlineExchangeCtor),
-		fx.Provide(onlineNamesysCtor),
-
-		fx.Invoke(ipnsRepublisher),
-
-		fx.Provide(p2p.NewP2P),
-
-		ipfsp2p,
-		providers,
-	)
-	if !cfg.Online {
-		online = fx.Options(
-			fx.Provide(offline.Exchange),
-			fx.Provide(offlineNamesysCtor),
-			fx.Provide(offroute.NewOfflineRouter),
-			fx.Provide(provider.NewOfflineProvider),
-		)
-	}
-
 	core := fx.Options(
-		fx.Provide(blockServiceCtor),
-		fx.Provide(dagCtor),
+		fx.Provide(node.BlockServiceCtor),
+		fx.Provide(node.DagCtor),
 		fx.Provide(resolver.NewBasicResolver),
-		fx.Provide(pinning),
-		fx.Provide(files),
+		fx.Provide(node.Pinning),
+		fx.Provide(node.Files),
 	)
 
 	n := &IpfsNode{
@@ -214,16 +133,16 @@ func NewNode(ctx context.Context, cfg *BuildCfg) (*IpfsNode, error) {
 	}
 
 	app := fx.New(
+		fx.NopLogger,
 		fx.Provide(baseProcess),
 
 		params,
-		storage,
-		ident,
-		ipns,
-		online,
+		node.Storage,
+		node.Identity,
+		node.IPNS,
+		node.Networked(cfg.Online),
 
 		fx.Invoke(setupSharding),
-		fx.NopLogger,
 
 		core,
 
@@ -248,19 +167,10 @@ func NewNode(ctx context.Context, cfg *BuildCfg) (*IpfsNode, error) {
 		return nil, err
 	}
 
-	// TODO: DI-ify bootstrap
+	// TODO: How soon will bootstrap move to libp2p?
 	if !cfg.Online {
 		return n, nil
 	}
 
-	return n, n.Bootstrap(DefaultBootstrapConfig)
-}
-
-func isTooManyFDError(err error) bool {
-	perr, ok := err.(*os.PathError)
-	if ok && perr.Err == syscall.EMFILE {
-		return true
-	}
-
-	return false
+	return n, n.Bootstrap(bootstrap.DefaultBootstrapConfig)
 }
