+++ test/integration/bitswap_wo_routing_test.go
@@ -8,6 +8,7 @@ import (
 	"github.com/ipfs/go-block-format"
 	"github.com/ipfs/go-ipfs/core"
 	"github.com/ipfs/go-ipfs/core/mock"
+	"github.com/ipfs/go-ipfs/core/node"
 
 	cid "github.com/ipfs/go-cid"
 	mocknet "github.com/libp2p/go-libp2p/p2p/net/mock"
@@ -26,7 +27,7 @@ func TestBitswapWithoutRouting(t *testing.T) {
 		n, err := core.NewNode(ctx, &core.BuildCfg{
 			Online:  true,
 			Host:    coremock.MockHostOption(mn),
-			Routing: core.NilRouterOption, // no routing
+			Routing: node.NilRouterOption, // no routing
 		})
 		if err != nil {
 			t.Fatal(err)
